pyproject.toml
[project]
name = "cirno-questions-generator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a]>=1.28.0",
    "httpx>=0.28.1",
    "litellm==1.82.6",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
